refactor(ffi/errors): rename `InternalError` -> `CoreCryptoError`
Here are the constraints we are operating under:

- Wasm requires a wrapper aroud the normal `CoreCryptoError` to get its
  specialty error handling working properly.
- Uniffi can't deal with an error wrapper at all.
- Uniffi can't rename its error type.
- Uniffi can't deal with an error called `InternalError` because it
  defines its own.
- Everything everywhere needs to return `CoreCryptoResult` which
  returns by default a `CoreCryptoError`.

This refactor navigates through all those constraints.
