+++ crypto-ffi/src/core_crypto/command/mod.rs
@@ -83,7 +83,7 @@ impl CoreCrypto {
 
                 // In wasm only, we are required to manually convert the error type. Uniffi does it for us.
                 #[cfg(target_family = "wasm")]
-                let err = crate::error::internal::InternalError::TransactionFailed(err).into();
+                let err = crate::error::core_crypto::CoreCryptoError::TransactionFailed(err).into();
 
                 Err(err)
             }
