+++ crypto-ffi/src/lib.rs
@@ -53,9 +53,7 @@ pub use e2ei::{
 };
 #[cfg(feature = "proteus")]
 pub use error::proteus::ProteusError;
-pub use error::{CoreCryptoResult, core_crypto::CoreCryptoError, mls::MlsError};
-#[cfg(target_family = "wasm")]
-pub use error::{WasmCryptoResult, internal::InternalError};
+pub use error::{CoreCryptoError, CoreCryptoResult, mls::MlsError};
 pub use identity::{
     wire::{DeviceStatus, WireIdentity},
     x509::X509Identity,
